auth_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
bcrypt==3.2.2
PyJWT==2.9.0
python-multipart==0.0.9
